Refactory folder exception
