+++ src/app.controller.ts
@@ -1,7 +1,5 @@
 import { Controller, Get, Post } from '@nestjs/common';
 import { ConfigService } from './config/config.service';
-import { JsonplaceholderService } from './jsonplaceholder/jsonplaceholder.service';
-import { UserService } from './user/user.service';
 
 @Controller('/app')
 export class AppController {
