+++ src/user/user.controller.ts
@@ -1,8 +1,10 @@
-import { Controller, Get, Param } from '@nestjs/common';
+import { Controller, Get, Param, UseFilters, NotFoundException } from '@nestjs/common';
 import { UserService } from './user.service';
 import { User } from './user.entity';
+import { UserExceptionFilter } from './user.exception.filter';
 
 @Controller('user')
+@UseFilters(new UserExceptionFilter())
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
@@ -13,6 +15,10 @@ export class UserController {
 
   @Get('/:id')
   async findOne(@Param('id') id): Promise<User> {
-    return await this.userService.findById(id);
+    const user = await this.userService.findById(id);
+    if (!user) {
+      throw new NotFoundException();
+    }
+    return user;
   }
 }
